fix(e2e): tree-expand fixes for mapping suggest (#315)
PR-B-AI's mapping-suggest.spec.ts inherited the same broken
openRequirementDetail helper that the Phase 3 agents wrote for each
new spec — it clicked the leaf row's text without first expanding
ancestor categories. The seeded catalogs (SOC 2 / ISO 27001 / HIPAA)
have ONLY category rows at the top level, so leaves are hidden behind
collapsed categories. Every test that walked through
openRequirementDetail timed out.

Extract the recursive leaf-discovery + ancestor-expanding logic into
frontend/e2e/_mapping-helpers.ts (same file added on
feat/mapping-history-ui) and have mapping-suggest.spec.ts import from
it. FrameworkDetail's kebab on this branch still has just 2 items so
the dropdown-overlap regression seen on PR-C1 and PR-D does not apply
here.
